#2168 Refactoring Device & EdgeDevice sync job for AWS
- 1 job Sync things job => update device & edgedevice based on thingtype (iotEdge or not)
- Remove Sync GreenGrass Devices
- Update the job thingtype for isEdge method
