+++ src/AzureIoTHub.Portal.Tests.Unit/Infrastructure/Jobs/AWS/SyncThingsJobTests.cs
@@ -9,11 +9,14 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
     using System.Net;
     using System.Threading;
     using System.Threading.Tasks;
+    using Amazon.GreengrassV2;
+    using Amazon.GreengrassV2.Model;
     using Amazon.IoT;
     using Amazon.IoT.Model;
     using Amazon.IotData;
     using Amazon.IotData.Model;
     using AutoFixture;
+    using AzureIoTHub.Portal.Application.Services.AWS;
     using AzureIoTHub.Portal.Domain;
     using AzureIoTHub.Portal.Domain.Entities;
     using AzureIoTHub.Portal.Domain.Repositories;
@@ -30,10 +33,14 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
         private Mock<IAmazonIoT> amazonIoTClient;
         private Mock<IAmazonIotData> amazonIoTDataClient;
+        private Mock<IAmazonGreengrassV2> amazonGreenGrass;
         private Mock<IUnitOfWork> mockUnitOfWork;
         private Mock<IDeviceRepository> mockDeviceRepository;
+        private Mock<IEdgeDeviceRepository> mockEdgeDeviceRepository;
         private Mock<IDeviceModelRepository> mockDeviceModelRepository;
+        private Mock<IEdgeDeviceModelRepository> mockEdgeDeviceModelRepository;
         private Mock<IDeviceTagValueRepository> mockDeviceTagValueRepository;
+        private Mock<IAWSExternalDeviceService> awsExternalDeviceService;
 
         public override void Setup()
         {
@@ -41,17 +48,25 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             this.mockUnitOfWork = MockRepository.Create<IUnitOfWork>();
             this.mockDeviceRepository = MockRepository.Create<IDeviceRepository>();
+            this.mockEdgeDeviceRepository = MockRepository.Create<IEdgeDeviceRepository>();
             this.mockDeviceModelRepository = MockRepository.Create<IDeviceModelRepository>();
+            this.mockEdgeDeviceModelRepository = MockRepository.Create<IEdgeDeviceModelRepository>();
             this.mockDeviceTagValueRepository = MockRepository.Create<IDeviceTagValueRepository>();
             this.amazonIoTClient = MockRepository.Create<IAmazonIoT>();
             this.amazonIoTDataClient = MockRepository.Create<IAmazonIotData>();
+            this.amazonGreenGrass = MockRepository.Create<IAmazonGreengrassV2>();
+            this.awsExternalDeviceService = MockRepository.Create<IAWSExternalDeviceService>();
 
             _ = ServiceCollection.AddSingleton(this.mockUnitOfWork.Object);
             _ = ServiceCollection.AddSingleton(this.mockDeviceRepository.Object);
+            _ = ServiceCollection.AddSingleton(this.mockEdgeDeviceRepository.Object);
             _ = ServiceCollection.AddSingleton(this.mockDeviceModelRepository.Object);
+            _ = ServiceCollection.AddSingleton(this.mockEdgeDeviceModelRepository.Object);
             _ = ServiceCollection.AddSingleton(this.mockDeviceTagValueRepository.Object);
             _ = ServiceCollection.AddSingleton(this.amazonIoTClient.Object);
             _ = ServiceCollection.AddSingleton(this.amazonIoTDataClient.Object);
+            _ = ServiceCollection.AddSingleton(this.amazonGreenGrass.Object);
+            _ = ServiceCollection.AddSingleton(this.awsExternalDeviceService.Object);
             _ = ServiceCollection.AddSingleton<IJob, SyncThingsJob>();
 
 
@@ -87,7 +102,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -99,6 +114,17 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                     Version = newDevice.Version,
                     HttpStatusCode = HttpStatusCode.OK
                 });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(false);
 
             _ = this.mockDeviceModelRepository
                 .Setup(x => x.GetByNameAsync(newDevice.DeviceModel.Name))
@@ -123,6 +149,17 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 });
 
             this.mockDeviceRepository.Setup(x => x.Delete(It.IsAny<string>())).Verifiable();
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>
+                {
+                    new EdgeDevice
+                    {
+                        Id = Guid.NewGuid().ToString()
+                    }
+                });
+
+            this.mockEdgeDeviceRepository.Setup(x => x.Delete(It.IsAny<string>())).Verifiable();
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -170,7 +207,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -182,6 +219,17 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                     Version = 2,
                     HttpStatusCode = HttpStatusCode.OK
                 });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(false);
 
             _ = this.mockDeviceModelRepository
                 .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
@@ -199,6 +247,9 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -237,7 +288,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -249,6 +300,17 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                     Version = 1,
                     HttpStatusCode = HttpStatusCode.OK
                 });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(false);
 
             _ = this.mockDeviceModelRepository
                 .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
@@ -262,6 +324,9 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -300,7 +365,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -311,6 +376,9 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -349,7 +417,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -364,6 +432,9 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
 
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
+
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
 
@@ -375,7 +446,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
         }
 
         [Test]
-        public async Task ExecuteNewDeviceWithUnknownThingTypeSkipped()
+        public async Task ExecuteNewDeviceWithUnknownIsEdgeTagSkipped()
         {
             // Arrange
             var mockJobExecutionContext = MockRepository.Create<IJobExecutionContext>();
@@ -401,7 +472,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -414,12 +485,93 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                     HttpStatusCode = HttpStatusCode.OK
                 });
 
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync((bool?)null);
+
+            _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
+
+            _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
+                .Returns(Task.CompletedTask);
+
+            // Act
+            await this.syncThingsJob.Execute(mockJobExecutionContext.Object);
+
+            // Assert
+            MockRepository.VerifyAll();
+        }
+
+        [Test]
+        public async Task ExecuteNewDeviceWithUnknownDeviceModelSkipped()
+        {
+            // Arrange
+            var mockJobExecutionContext = MockRepository.Create<IJobExecutionContext>();
+
+            var expectedDeviceModel = Fixture.Create<DeviceModel>();
+            var existingDevice = new Device
+            {
+                Id = Fixture.Create<string>(),
+                Name = Fixture.Create<string>(),
+                DeviceModel = expectedDeviceModel,
+                DeviceModelId = expectedDeviceModel.Id,
+                Version = 2
+            };
+
+            var thingsListing = new ListThingsResponse
+            {
+                Things = new List<ThingAttribute>()
+                {
+                    new ThingAttribute
+                    {
+                        ThingName = existingDevice.Name
+                    }
+                }
+            };
+
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(thingsListing);
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingResponse()
+                {
+                    ThingId = existingDevice.Id,
+                    ThingName = existingDevice.Name,
+                    ThingTypeName = existingDevice.DeviceModel.Name,
+                    Version = 1,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(false);
+
             _ = this.mockDeviceModelRepository
                 .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
                 .ReturnsAsync((DeviceModel)null);
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -459,7 +611,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -472,6 +624,17 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                     HttpStatusCode = HttpStatusCode.OK
                 });
 
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(false);
+
             _ = this.mockDeviceModelRepository
                 .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
                 .ReturnsAsync(expectedDeviceModel);
@@ -481,6 +644,9 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -518,7 +684,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -526,6 +692,9 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
@@ -563,7 +732,7 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                 }
             };
 
-            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<CancellationToken>()))
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(thingsListing);
 
             _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
@@ -576,6 +745,17 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
                     HttpStatusCode = HttpStatusCode.OK
                 });
 
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(false);
+
             _ = this.mockDeviceModelRepository
                 .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
                 .ReturnsAsync(expectedDeviceModel);
@@ -585,6 +765,343 @@ namespace AzureIoTHub.Portal.Tests.Unit.Infrastructure.Jobs.AWS
 
             _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
                 .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
+
+            _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
+                .Returns(Task.CompletedTask);
+
+            // Act
+            await this.syncThingsJob.Execute(mockJobExecutionContext.Object);
+
+            // Assert
+            MockRepository.VerifyAll();
+        }
+
+        [Test]
+        public async Task ExecuteNewEdgeDeviceEdgeDeviceCreated()
+        {
+            // Arrange
+            var mockJobExecutionContext = MockRepository.Create<IJobExecutionContext>();
+
+            var expectedDeviceModel = Fixture.Create<EdgeDeviceModel>();
+            var newDevice = new EdgeDevice
+            {
+                Id = Fixture.Create<string>(),
+                Name = Fixture.Create<string>(),
+                DeviceModel = expectedDeviceModel,
+                DeviceModelId = expectedDeviceModel.Id,
+                Version = 1
+            };
+
+            var thingsListing = new ListThingsResponse
+            {
+                Things = new List<ThingAttribute>()
+                {
+                    new ThingAttribute
+                    {
+                        ThingName = newDevice.Name
+                    }
+                }
+            };
+
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(thingsListing);
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingResponse()
+                {
+                    ThingId = newDevice.Id,
+                    ThingName = newDevice.Name,
+                    ThingTypeName = newDevice.DeviceModel.Name,
+                    Version = newDevice.Version,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(true);
+
+            _ = this.mockEdgeDeviceModelRepository
+                .Setup(x => x.GetByNameAsync(newDevice.DeviceModel.Name))
+                .ReturnsAsync(expectedDeviceModel);
+
+            _ = this.amazonGreenGrass.Setup(client => client.GetCoreDeviceAsync(It.IsAny<GetCoreDeviceRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new GetCoreDeviceResponse() { HttpStatusCode = HttpStatusCode.OK });
+
+            _ = this.mockEdgeDeviceRepository.Setup(repository => repository.GetByIdAsync(newDevice.Id, d => d.Tags))
+                .ReturnsAsync((EdgeDevice)null);
+
+            _ = this.mockEdgeDeviceRepository.Setup(repository => repository.InsertAsync(It.IsAny<EdgeDevice>()))
+                .Returns(Task.CompletedTask);
+
+            _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<Device>
+                {
+                    new Device
+                    {
+                        Id = Guid.NewGuid().ToString()
+                    }
+                });
+
+            this.mockDeviceRepository.Setup(x => x.Delete(It.IsAny<string>())).Verifiable();
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>
+                {
+                    new EdgeDevice
+                    {
+                        Id = Guid.NewGuid().ToString()
+                    }
+                });
+
+            this.mockEdgeDeviceRepository.Setup(x => x.Delete(It.IsAny<string>())).Verifiable();
+
+            _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
+                .Returns(Task.CompletedTask);
+
+            // Act
+            await this.syncThingsJob.Execute(mockJobExecutionContext.Object);
+
+            // Assert
+            MockRepository.VerifyAll();
+        }
+
+        [Test]
+        public async Task ExecuteExistingEdgeDeviceWithHigherVersionEdgeDeviceUpdated()
+        {
+            // Arrange
+            var mockJobExecutionContext = MockRepository.Create<IJobExecutionContext>();
+
+            var expectedDeviceModel = Fixture.Create<EdgeDeviceModel>();
+            var existingDevice = new EdgeDevice
+            {
+                Id = Fixture.Create<string>(),
+                Name = Fixture.Create<string>(),
+                DeviceModel = expectedDeviceModel,
+                DeviceModelId = expectedDeviceModel.Id,
+                Version = 1,
+                Tags = new List<DeviceTagValue>()
+                {
+                    new()
+                    {
+                        Id = Fixture.Create<string>(),
+                        Name = Fixture.Create<string>(),
+                        Value = Fixture.Create<string>()
+                    }
+                }
+            };
+
+            var thingsListing = new ListThingsResponse
+            {
+                Things = new List<ThingAttribute>()
+                {
+                    new ThingAttribute
+                    {
+                        ThingName = existingDevice.Name
+                    }
+                }
+            };
+
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(thingsListing);
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingResponse()
+                {
+                    ThingId = existingDevice.Id,
+                    ThingName = existingDevice.Name,
+                    ThingTypeName = existingDevice.DeviceModel.Name,
+                    Version = 2,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(true);
+
+            _ = this.mockEdgeDeviceModelRepository
+                .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
+                .ReturnsAsync(expectedDeviceModel);
+
+            _ = this.amazonGreenGrass.Setup(client => client.GetCoreDeviceAsync(It.IsAny<GetCoreDeviceRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new GetCoreDeviceResponse() { HttpStatusCode = HttpStatusCode.OK });
+
+            _ = this.mockEdgeDeviceRepository.Setup(repository => repository.GetByIdAsync(existingDevice.Id, d => d.Tags))
+                .ReturnsAsync(existingDevice);
+
+            this.mockDeviceTagValueRepository.Setup(repository => repository.Delete(It.IsAny<string>())).Verifiable();
+
+            this.mockEdgeDeviceRepository.Setup(repository => repository.Update(It.IsAny<EdgeDevice>())).Verifiable();
+
+            _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
+
+            _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
+                .Returns(Task.CompletedTask);
+
+            // Act
+            await this.syncThingsJob.Execute(mockJobExecutionContext.Object);
+
+            // Assert
+            MockRepository.VerifyAll();
+        }
+
+        [Test]
+        public async Task ExecuteExistingEdgeDeviceWithOlderVersionEdgeDeviceNotUpdated()
+        {
+            // Arrange
+            var mockJobExecutionContext = MockRepository.Create<IJobExecutionContext>();
+
+            var expectedDeviceModel = Fixture.Create<EdgeDeviceModel>();
+            var existingDevice = new EdgeDevice
+            {
+                Id = Fixture.Create<string>(),
+                Name = Fixture.Create<string>(),
+                DeviceModel = expectedDeviceModel,
+                DeviceModelId = expectedDeviceModel.Id,
+                Version = 2
+            };
+
+            var thingsListing = new ListThingsResponse
+            {
+                Things = new List<ThingAttribute>()
+                {
+                    new ThingAttribute
+                    {
+                        ThingName = existingDevice.Name
+                    }
+                }
+            };
+
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(thingsListing);
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingResponse()
+                {
+                    ThingId = existingDevice.Id,
+                    ThingName = existingDevice.Name,
+                    ThingTypeName = existingDevice.DeviceModel.Name,
+                    Version = 1,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(true);
+
+            _ = this.mockEdgeDeviceModelRepository
+                .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
+                .ReturnsAsync(expectedDeviceModel);
+
+            _ = this.amazonGreenGrass.Setup(client => client.GetCoreDeviceAsync(It.IsAny<GetCoreDeviceRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new GetCoreDeviceResponse() { HttpStatusCode = HttpStatusCode.OK });
+
+            _ = this.mockEdgeDeviceRepository.Setup(repository => repository.GetByIdAsync(existingDevice.Id, d => d.Tags))
+                .ReturnsAsync(existingDevice);
+
+            _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
+
+            _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
+                .Returns(Task.CompletedTask);
+
+            // Act
+            await this.syncThingsJob.Execute(mockJobExecutionContext.Object);
+
+            // Assert
+            MockRepository.VerifyAll();
+        }
+
+        [Test]
+        public async Task ExecuteNewEdgeDeviceWithUnknownEdgeDeviceModelSkipped()
+        {
+            // Arrange
+            var mockJobExecutionContext = MockRepository.Create<IJobExecutionContext>();
+
+            var expectedDeviceModel = Fixture.Create<EdgeDeviceModel>();
+            var existingDevice = new EdgeDevice
+            {
+                Id = Fixture.Create<string>(),
+                Name = Fixture.Create<string>(),
+                DeviceModel = expectedDeviceModel,
+                DeviceModelId = expectedDeviceModel.Id,
+                Version = 2
+            };
+
+            var thingsListing = new ListThingsResponse
+            {
+                Things = new List<ThingAttribute>()
+                {
+                    new ThingAttribute
+                    {
+                        ThingName = existingDevice.Name
+                    }
+                }
+            };
+
+            _ = this.amazonIoTClient.Setup(client => client.ListThingsAsync(It.IsAny<ListThingsRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(thingsListing);
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingAsync(It.IsAny<DescribeThingRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingResponse()
+                {
+                    ThingId = existingDevice.Id,
+                    ThingName = existingDevice.Name,
+                    ThingTypeName = existingDevice.DeviceModel.Name,
+                    Version = 1,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.amazonIoTClient.Setup(client => client.DescribeThingTypeAsync(It.IsAny<DescribeThingTypeRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new DescribeThingTypeResponse()
+                {
+                    ThingTypeId = expectedDeviceModel.Id,
+                    ThingTypeName = expectedDeviceModel.Name,
+                    HttpStatusCode = HttpStatusCode.OK
+                });
+
+            _ = this.awsExternalDeviceService.Setup(client => client.IsEdgeThingType(It.IsAny<DescribeThingTypeResponse>()))
+                .ReturnsAsync(true);
+
+            _ = this.mockEdgeDeviceModelRepository
+                .Setup(x => x.GetByNameAsync(existingDevice.DeviceModel.Name))
+                .ReturnsAsync((EdgeDeviceModel)null);
+
+            _ = this.mockDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<Device, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<Device>());
+
+            _ = this.mockEdgeDeviceRepository.Setup(x => x.GetAllAsync(It.IsAny<Expression<Func<EdgeDevice, bool>>>(), It.IsAny<CancellationToken>(), d => d.Tags, d => d.Labels))
+                .ReturnsAsync(new List<EdgeDevice>());
 
             _ = this.mockUnitOfWork.Setup(work => work.SaveAsync())
                 .Returns(Task.CompletedTask);
